# api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Qeet Pay API
  description: >-
    Qeet Pay — India-first payments, billing and GST infrastructure (UPI / Cards / NACH). Programmatic `/v1` calls
    authenticate with a `qp_live_` / `qp_test_` API key via the `X-Api-Key` header; money is always in integer minor
    units (paise).
  contact:
    name: Qeet Group
    url: https://qeet.in
  license:
    name: Proprietary
    url: https://qeet.in
  version: v1
servers:
  - url: https://api.pay.qeet.in
    description: Production
  - url: https://api.pay.staging.qeet.in
    description: Staging
  - url: http://localhost:4201
    description: Local (docker compose)
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/payouts:
    post:
      tags:
        - Payouts
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayoutRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PayoutView'
  /v1/payouts/{id}/reject:
    post:
      tags:
        - Payouts
      operationId: reject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PayoutView'
  /v1/payouts/{id}/approve:
    post:
      tags:
        - Payouts
      operationId: approve
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /v1/payout-batches:
    get:
      tags:
        - Payouts
      operationId: list_5
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PayoutBatchSummaryView'
    post:
      tags:
        - Payouts
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PayoutBatchView'
  /v1/payout-batches/{id}/reject:
    post:
      tags:
        - Payouts
      operationId: reject_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PayoutBatchView'
  /v1/payout-batches/{id}/approve:
    post:
      tags:
        - Payouts
      operationId: approve_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /v1/payout-batches/csv:
    post:
      tags:
        - Payouts
      operationId: createFromCsv
      parameters:
        - name: currency
          in: query
          required: false
          schema:
            type: string
            default: INR
        - name: description
          in: query
          required: false
          schema:
            type: string
      requestBody:
        content:
          text/csv:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PayoutBatchView'
...
  /v1/insurance/claims/{claimId}/approve:
    post:
      tags:
        - Insurance
      operationId: approve_2
      parameters:
        - name: claimId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ClaimView'
...
  /v1/payouts/{id}:
    get:
      tags:
        - Payouts
      operationId: get_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PayoutView'
  /v1/payout-batches/{id}:
    get:
      tags:
        - Payouts
      operationId: get_6
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PayoutBatchView'
...
components:
  schemas:
...
    CreatePayoutRequest:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
          minLength: 1
        rail:
          type: string
          enum:
            - UPI
            - IMPS
            - NEFT
            - RTGS
        destination:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - currency
        - destination
        - rail
    PayoutView:
      type: object
      properties:
        id:
          type: string
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
        rail:
          type: string
        status:
          type: string
        ledgerEntryId:
          type: string
    CreateBatchRequest:
      type: object
      properties:
        currency:
          type: string
          minLength: 1
        description:
          type: string
        payouts:
          type: array
          items:
            $ref: '#/components/schemas/PayoutLineDto'
          minItems: 1
      required:
        - currency
        - payouts
    PayoutLineDto:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        rail:
          type: string
          enum:
            - UPI
            - IMPS
            - NEFT
            - RTGS
        destination:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - destination
        - rail
    PayoutBatchView:
      type: object
      properties:
        id:
          type: string
        currency:
          type: string
        status:
          type: string
        totalCount:
          type: integer
          format: int32
        totalAmountMinor:
          type: integer
          format: int64
        paidCount:
          type: integer
          format: int32
        failedCount:
          type: integer
          format: int32
        description:
          type: string
        payouts:
          type: array
          items:
            $ref: '#/components/schemas/PayoutLineView'
    PayoutLineView:
      type: object
      properties:
        id:
          type: string
        amountMinor:
          type: integer
          format: int64
        rail:
          type: string
        destination:
          type: string
        status:
          type: string
        ledgerEntryId:
          type: string
...
    ClaimView:
      type: object
      properties:
        id:
          type: string
        policyId:
          type: string
        amountMinor:
          type: integer
          format: int64
        reason:
          type: string
        status:
          type: string
        payoutEntryId:
          type: string
        createdAt:
          type: string
          format: date-time
        decidedAt:
          type: string
          format: date-time
...
    PayoutBatchSummaryView:
      type: object
      properties:
        id:
          type: string
        currency:
          type: string
        status:
          type: string
        totalCount:
          type: integer
          format: int32
        totalAmountMinor:
          type: integer
          format: int64
        paidCount:
          type: integer
          format: int32
        failedCount:
          type: integer
          format: int32
        description:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Programmatic API key (`qp_live_…` / `qp_test_…`) presented as the `X-Api-Key` header.
      name: X-Api-Key
      in: header
...